# siron-activate-main 2/apis/v1/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: Informatique-MTF S.A.
  description: Provide functionality for feature licensing management
  title: Siron Activate REST API
  version: v1.4.0
servers:
  - url: https://siron-activate.imtf-group.com
    description: Production server (uses live data)
  - url: https://licensing-service.vmo-dev-licensing.dev.imtf-group.com
    description: Sandbox server (uses test data)
...
security:
  - openId:
      - user
      - license-issuer
      - license-user
      - admin
paths:
  /licensing/api/v1/capabilities:
    post:
      operationId: registerCapability
      tags:
        - capabilities
      summary: Register a capability and its optional corresponding features.
      security:
        - openId:
            - admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Capability'
            examples:
              Capability static example without features:
                value:
                  name: acm
              Capability static example:
                value:
                  name: acm
                  features:
                    - name: Feature A
                    - name: Feature B

        description: The capability with its optional features to register

      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capability'
              examples:
                Capability static example without feature:
                  value:
                    oid: 694bfe51-1731-43aa-8fef-670e3f249c0a
                    name: 'acm'
                Capability static example:
                  value:
                    oid: ca88e94e-e9f2-4c3d-9ae0-ae5b029c13d3
                    name: acm
                    features:
                      - name: Feature A
                      - name: Feature B

          description: The successfully registered capability
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error when validating the request
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL not found

    get:
      operationId: getRegisteredCapabilities
      tags:
        - capabilities
      summary: Return a list of all registered capabilities and their corresponding features.
      security:
        - openId:
            - user
            - license-issuer
            - admin
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Capability'
              examples:
                Capability static example without feature:
                  value:
                    - oid: daa26016-41eb-4b64-bfd3-3cbf48194dd8
                      name: acm
                    - oid: 8ece7ce2-2266-41cc-9cc6-4d374bbd3448
                      name: detect
                Capability static example:
                  value:
                    - oid: 699c826f-f342-41e2-a080-afaa7f620c22
                      name: acm
                      features:
                        - name: Feature A
                        - name: Feature B
                    - oid: 43969a29-db0e-4916-b661-a9dda10f85f0
                      name: detect
                      features:
                        - name: Feature A


          description: The registered capabilities
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL not found
  /licensing/api/v1/capabilities/{oid}:
    put:
      operationId: updateCapability
      tags:
        - capabilities
      summary: Update an capability given by its oid (override).
      security:
        - openId:
            - admin
      parameters:
        - in: path
          name: oid
          schema:
            type: string
          required: true
          description: the oid of the capability to update

      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Capability'
            examples:
              Capability static example wihtout features:
                value:
                  oid: 61142af3-a38b-46f7-9b1f-50a9f6d0e4b1
                  name: acm
              Capability static example:
                value:
                  oid: 01658bfd-2e8c-4acc-981c-a44c312a4ff2
                  name: acm
                  features:
                    - name: Feature A
                    - name: Feature B

        description: The capability to update. The oid of the capability request body will be ignored and overriden by the one provided as path parameter.

      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capability'
              examples:
                Capability static example without feature:
                  value:
                    oid: ad4d047e-e301-4daa-a2f9-f6d0600219e9
                    name: 'acm'
                Capability static example:
                  value:
                    oid: ea7b8550-840f-49e4-8d8e-c41f774bdc74
                    name: acm
                    features:
                      - name: Feature A
                      - name: Feature B

          description: The successfully registered capability
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error when validating the request
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL or capability oid not found

    delete:
      operationId: deleteCapability
      tags:
        - capabilities
      summary: Delete an capability given by its oid.
      security:
        - openId:
            - admin
      parameters:
        - in: path
          name: oid
          schema:
            type: string
          required: true
          description: the oid of the capability to update

      responses:
        "204":
          description: Capability successfully deleted
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL or capability oid not found
  /licensing/api/v1/customers:
    post:
      operationId: addCustomer
      tags:
        - customers
      summary: Add a new customer
      security:
        - openId:
            - license-issuer
            - admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
            examples:
              Static example:
                value:
                  name: John Doe
        description: the customer to add
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
              examples:
                Static example:
                  value:
                    oid: 37375334-1066-4b2e-88a8-814c59c4010f
                    name: John Doe
          description: The registered customer
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL not found

    get:
      operationId: getCustomers
      tags:
        - customers
      summary: Return the list of all customers
      security:
        - openId:
            - user
            - license-issuer
            - admin
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
              examples:
                Static example:
                  value:
                    - oid: 24accaf8-c274-4041-ad14-c3a5b48c6327
                      name: John Doe
                    - oid: 6e761ba6-37a5-480b-9455-8b4a2ed916af
                      name: Jane Doe

          description: The list of all customers
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL not found
  /licensing/api/v1/customers/{oid}:
    put:
      operationId: updateCustomer
      tags:
        - customers
      summary: Update an existing customer given by its oid (override)
      security:
        - openId:
            - admin
      parameters:
        - in: path
          name: oid
          schema:
            type: string
          required: true
          description: the oid of the customer to update
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
            examples:
              Static example:
                value:
                  oid: 8315f9c6-3383-45bf-8154-2c06cee5ddb8
                  name: John Doe

        description: the customer to update
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
              examples:
                Static example:
                  value:
                    oid: b813ac63-29be-494b-88af-7e4d5a967e9c
                    name: John Doe

          description: The updated customer
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL or customer oid not found

    delete:
      operationId: deleteCustomer
      tags:
        - customers
      summary: Delete an existing customer given by its oid
      security:
        - openId:
            - admin
      parameters:
        - in: path
          name: oid
          schema:
            type: string
          required: true
          description: the oid of the customer to update
      responses:
        "204":
          description: Customer successfully deleted
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL or customer oid not found
...
  /licensing/api/v1/json/load:
    post:
      operationId: jsonImport
      tags:
        - import-export
      summary: Load licenses as JSON. This API will override all licenses stored in the system.
      security:
        - openId:
            - admin
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/License'
            examples:
              Static example:
                value:
                  - oid: 6ca6280b-a6fb-4a78-aba8-03933b190f2e
                    capabilities: ["acm", "aml"]
                    productName: acm
                    customerId: johnDoe1234
                    customerName: John Doe
                    creationDate: "2023-02-15T14:30:00.000Z"
                    expirationDate: "2025-03-01"
                    contactEmail: v@t.com
                  - oid: 8dd8bfc3-4c23-40ba-86a1-c7899f9265c8
                    capabilities: ["ccm", "aml"]
                    productName: ""
                    customerId: johnDoe1234
                    customerName: John Doe
                    creationDate: "2023-02-15T14:30:00.000Z"
                    expirationDate: "2025-03-01"
                    contactEmail: v@t.com

        description: The licenses to import
      responses:
        "204":
          description: Licenses imported successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Error when validating the request
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL not found
  /licensing/api/v1/json/download:
    get:
      operationId: jsonExport
      tags:
        - import-export
      summary: Download all licenses stored in the system as JSON.
      security:
        - openId:
            - admin
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/License'
              examples:
                Static example:
                  value:
                    - oid: 6ca6280b-a6fb-4a78-aba8-03933b190f2e
                      capabilities: ["acm", "aml"]
                      productName: acm
                      customerId: johnDoe1234
                      customerName: John Doe
                      creationDate: "2023-02-15T14:30:00.000Z"
                      expirationDate: "2025-03-01"
                      contactEmail: v@t.com
                    - oid: 8dd8bfc3-4c23-40ba-86a1-c7899f9265c8
                      capabilities: ["ccm", "aml"]
                      productName: ""
                      customerId: johnDoe1234
                      customerName: John Doe
                      creationDate: "2023-02-15T14:30:00.000Z"
                      expirationDate: "2025-03-01"
                      contactEmail: v@t.com


          description: A list of licenses
        "401":
          description: Error when authenticating
        "403":
          description: User is not authorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: URL not found
...
components:
  securitySchemes:
    openId:
      type: openIdConnect
      openIdConnectUrl: https://keycloak.imtf-group.com/realms/imtf/.well-known/openid-configuration
      description: oidc connected to imtf internal keycloak
  schemas:
    Capability:
      type: object
      description: An capability's definition registered
      properties:
        oid:
          type: string
          description: the technical uuid assigned to the capability
        name:
          type: string
          description: The name of the capability
        features:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                description: feature name
            required:
              - name
      required:
        - name
    Customer:
      type: object
      description: A customer representation
      properties:
        oid:
          type: string
          description: the technical uuid assigned to the customer
        name:
          type: string
          description: The name of the customer
      required:
        - name
    License:
      type: object
      description: the license definition
      properties:
        oid:
          type: string
          description: the technical uuid assigned to the license
        capabilities:
          type: array
          description: a list of capability names for which the license has been created.
          items:
            type: string
        productName:
          type: string
          description: the name of the product for which the license has been created, it needs to match an already registered capability.
        customerId:
          type: string
          description: a string id representing the customer for which the license has been created.
        customerName:
          type: string
          description: the display name of the customer for which the license has been created. Usually, the value of this field is displayed in the capability using the license.
        creationDate:
          type: string
          format: date-time
          description: the creation date of the license
        expirationDate:
          type: string
          format: date
          description: the expiration date of the license
        signatureDate:
          type: string
          format: date-time
          description: the date of the first signature of the license (deprecated since 1.1.0)
          deprecated: true
        signedBy:
          type: string
          description: the name of the user who first signed the license (deprecated since 1.1.0)
          deprecated: true
        contactEmail:
          type: string
          format: email
          description: contact information, email or name, about the owner of this license
      required:
        - productName
        - customerId
        - customerName
        - creationDate
...
    ErrorResponse:
      type: object
      description: Error is used to return the error details inside error response
        object
      properties:
        code:
          type: string
          description: Unique identifier of the error type
        message:
          type: string
          description: Simple message describing the error
        timestamp:
          type: string
          format: date-time
          description: Exact time of the error, UTC based with ISO-8601 format
        spanId:
          type: string
          description: If available, the OpenTracing span Id
        traceId:
          type: string
          description: If available, the OpenTracing trace Id
      required:
        - code
        - message
        - timestamp
...